requirements.txt
openai
discord.py==2.2.2